move immutable graph.
